pyproject.toml
[project]
name = "civicarena"
version = "0.1.0"
description = "Multi-agent civic deliberation platform where AI citizen-agents debate policy issues using structured democratic protocols"
readme = "CONCEPT.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "rich>=13.7",
    "httpx>=0.27",
    "pydantic>=2.6",
    "ddgs>=9.0",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.3",
]

[project.scripts]
civicarena = "civicarena.cli:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
